Updated openstack/openstack Project: openstack/nova  cc3fd7cb9860702ee3b41f779f543d69deaa34e5
On port update check port binding worked

When updating a port, we should do the same port binding check that we
do when creating a port.

Although in the case of update, we can use the existing clean up code
that ensures we only delete the port if it was created by Nova.

blueprint prep-for-network-aware-scheduling

Change-Id: I3ac06f45305aea5e0c18ba777331fb2f233523f5