Updated openstack/openstack Project: openstack-infra/project-config  130b1bc7ea90faf4480908899cb137d3a59585c9
Make trove tempest job nv until plugin is added

Tempest recently removed it's tests for trove in preparation for the
trove plugin being completed. This commit makes the tempest trove job
non voting until it's migrated to using the plugin so as to not block
trove development in the meantime.

Change-Id: I8f896fb77dc1e3bd58e8b7905958c6e505d22627