Updated openstack/openstack Project: openstack/tempest  aaff550edab124a744b2888bcf7a7d70d3abfa0c
Fix search disk name for the config_drive in scenario test

This commit fixes search for config_drive device name in
test "verify_metadata_on_config_drive", that supports only lowercase.
It needs because "mkfs" tool, for example in RHEL, capitalize the label
for config_drive ("CONFIG-2" instead of "config-2"), that used
in test as filter for blkid.

Closes-Bug: #1596868
Change-Id: I09684a7a295ea12ac372aadac926fe34708a66d4