Updated openstack/openstack Project: openstack/diskimage-builder  867bfaa44d52977e3a4b294403363f511bf52368
Install docker for tests

We have some test cases which attempt to build docker images, therefore
we need docker.

Fix a few bugs that showed up when we run docker tests - we need to
docker rm with sudo and docker images don't always have a /tmp so check
before unmounting it.

Change-Id: I147d0ef3f2ea83f35bac568214573a6bde0b1967