Updated openstack/openstack Project: openstack-infra/project-config  28e12113dd292dc5b0d01a98b89e296404745d66
Remove firefox, xvfb from bindep-fallback

We now have repositories with other-requirements that include these
packages, so the files will be cached on our daily builds.

Currently, we install these packages in jobs where they are needed, so
there is no need to install these on each and every non-devstack build.

Change-Id: Ia51738950b4d3e2e2b82107d437fc167cb9e1282