Updated openstack/openstack Project: openstack/oslo.db  581e1d13dc6643b4cb2a77a4573c7b77c298acb4
Make it possible to use enginefacade decorators with class methods

The decorator form can now be used with bound methods after a fix
for the related bug was merged. Mention this in docs and release
notes, so that people are aware of it.

A new test is added to make sure this also works with class methods
(as well as to check that the proposed decorator applying order
is actually correct).

Related-Bug: #1520195

Change-Id: Ifea08114d6d89de9d67fcae397eb94c0afc4d339