Updated openstack/openstack Project: openstack-infra/zuul  92dc6fbcd55f6145ebb684a1060330768610a953
Ansible launcher: clean up result handling

Have runJob always return a string or None for simplicity.
Construct the result dictionary only when sending the work_complete
packet, and send a dictionary with a result element every time, even
if the value is None.  Zuul will react the same whether there is
no dictionary, or it has no result or the result is null, and retry
the job.  Finally, send a simple string as status rather than the
result dictionary to ZMQ.  If the job is aborted, status will be
null.

Change-Id: Ie093c070c2fb4baf67f538b79826a7d8e42b18d9