Updated openstack/openstack Project: openstack/trove  21caa1373b4dac8a789cb37f1981f08b3adb2606
Add trove tempest plugin

This commit takes the trove tests which previously lived in tempest
and creates an in-tree tempest plugin out of them. The contents of
the plugin are taken with tempest checkout out at commit:
eda607cb1c89762a54290c4525d81c1b192c421f with modifications made for
things to live in the trove repo.

Change-Id: Ibad5502cdca0e34cca4b5bc8a7106611f7393cf2