Updated openstack/openstack Project: openstack-infra/project-config  b5c4058c7b049e13fba581e4abc36da8f991dd65
Fix doc publishing for infra

Documents that we only publish in the release pipeline, need to get
published to the top level directory as well as to a versioned subdir.

Create docs-tags job for publishing of these repositories, like it was
done for normal python projects.

Remove the storyboard publish jobs, we can use the normal infra publish
jobs for these. ci.openstack.org is not used anymore and the location
http://ci.openstack.org/storyboard
redirects to http://docs.openstack.org/infra/storyboard/

Also, update a misleading description.

Change-Id: Ibef97f8d0c06b9f7f000872de14283e2911d3099