Updated openstack/openstack Project: openstack/tempest  ff4f58862842c37f27788278c2bbce13ef2367e2
Add plugin registry generation to sphinx build

This commit adds running the plugin registry generation script to the
sphinx build. This ensure that every time we build the docs we'll have
an up-to-date plugin registry. This comes at the cost of increased
build time since the script will query openstack's git server to get
the plugin list. Also, the plugin-registry source file was added to
.gitignore since it is autogenerated now.

Change-Id: Ib26124c5d5d5ef4617e4e702e594ce7f1795ce23