Updated openstack/openstack Project: openstack/nova  222a88a54ceda741e0409f86aeb632b38a1792be
Always call port_update in allocate_for_instance

When we pull port_create out of allocate_for_instance, this will mean we
will always need to call port_update in allocate_for_instance so it can
add all the extra details we get from the virt driver.

We want to move most of the updates to the update phase of
allocate_for_instance. The port create just needs to do the minimal job
of creating the requested port. As such, we will now always call update.

blueprint prep-for-network-aware-scheduling

Change-Id: I4f6ad6c5c8e8bbc434672f5c12ebddd7afdff190