Updated openstack/openstack Project: openstack/heat  84791470809023d1e432522a6d7ea6c462b15c3c
Add context to stack lock function arguments

Context is added as the first argument to stack_lock_* db functions
but is currently ignored for getting the session used for stack lock
operations.

This is required later for bug 1479723 but is added in its own change
here to ease reviewing load.

Change-Id: Ieb3e4e2ee67150777cbe1e961d0d1806cf1f7e46
Related-Bug: #1479723