Updated openstack/openstack Project: openstack/neutron  1c4289e27f2d939d0803d8d1ba0f4eb464cc79e0
Add two more callbacks registry events

Change 877778ee4c7 clarified the objective of the callback
registry events, and it introduced two new API related events.

This change introduces a new set of events that are meant
to be used for handling process-related operations, in fact
even though AFTER_CREATE was fit for the job, it looks somewhat
misused now.

Change-Id: I9e717e9e4ae4414067477712ac02870fe1041e70