Updated openstack/openstack Project: openstack/horizon  a452f499be4213a58ac9d5916cc9513604f9a879
Add a condition whether displaying the link jump to Encryption page

At the moment, the link jump to Volume Encryption page is enable
even though volume does not encrypted. In this case, this page
is only displayed "Volume is Unencrypted" and it is not informative.
Therefore, this patch add a condition whether displaying the link
jump to Volume Encryption page.

Change-Id: Ic915dc5f05316991ed6943f66db1d1def54dcbaf
Closes-bug: #1558056