Updated openstack/openstack Project: openstack/requirements  d12bad9b2782ebed1ab277012334ec9f179f6015
Add python-tackerclient to global requirements

python-tackerclient is required for tacker-service and for
other projects like Mistral workflows to integrate with tacker.
Currently following Mistral blueprint is blocked due to this,

https://blueprints.launchpad.net/mistral/+spec/tacker-api-actions

1) Is the library actively maintained?

Yes, see https://launchpad.net/python-tackerclient

2) Is the library good code?

Yes. We have unit test running in gate to sanitize code quality

3) Is the library python 3 compatible?

No. While we run py34 job (non-voting) in the gate, we have few
failures that we plan to fix in Newton release.

4) Is the library license compatible?

Yes, python-tackerclient is licensed under Apache License, v2.0

5) Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

Not yet, we plan to work with the distros to package tacker in the Newton release

6) Is the function of this library already covered by other libraries in global-requirements.txt?

No

7) Is the library required for OpenStack project or related dev or infrastructure setup?

Yes, OpenStack Tacker

Change-Id: I206b7fa5268601f26adcacc5730a389ac3c90342