Updated openstack/openstack Project: openstack/nova  bc13221561ded371335248149143e14f40e32e99
Make VIF.address unique with port id for neutron

Neutron does not require global uniqueness for MAC addresses, but
nova-network did. Since VirtualInterface was designed for n-net and
is still used to enforce its requirements for such uniqueness, we can
simply use the extra space in the address field to namespace our
MAC address in neutron to avoid the constraint. We can hide this
behavior in the object so nothing else needs to know, and we have
255 characters in the field, which is enough for a uuid and a mac
address to fit.

Closes-Bug: #1597770
Change-Id: I23480d98ca339c119100e8ab7df633ed8b80b1e9