Updated openstack/openstack Project: openstack/manila  6f506a87052e315f5f14c1edf6d592d782204b18
HPE3PAR: Handle exceptions on deleted shares

Fix to ensure that an attempt to delete a manila share will not
fail because the 3PAR filestore does not exist. When the filestore
does not exist, that means the share is already gone (or never was
created). Avoid the delete and all filestore use in this case to
prevent failures during delete.

Also catch and log exceptions during the post-delete cleanup.

Change-Id: I0e477d3dae5706547354bd029b40e3308349fabb
Closes-Bug: #1597940