Updated openstack/openstack Project: openstack-infra/system-config  24f623218311f9ad379b7063f1d8ac430f622407
Do not use async with zl restarter playbook

Running this playbook on the puppetmaster we consistently run into ssh
failures due to async reconnecting periodically and network issues
between hosts. We can address this by starting a single connection
without async and polling on that which appears to be the default
wait_for behavior. Testing of this seems to indicate it is more
reliable.

Change-Id: Iec72e2c0d099c0e28bc4b4b48608a03b3e66b4c0