Updated openstack/openstack Project: openstack/diskimage-builder  5a5529fb84a2496f3fdaddf69be738219eb74d25
Fix the bug that "mktemp: failed to create directory"

In the function "run_in_target", it is failed that force an empty
TMPDIR inside the chroot. The TMPDIR is still the user defined
tmp dir. Due to the TMPDIR dir is not exist, using "mktemp" to
create tmp dir is failed.

Change-Id: I898f80099bc2a7c32e8676014d0f4263807f7039
Closes-Bug: #1597569