Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  fda36c1e2a32be7a9899b610956e28512a432db2
Add support for Clang Scan-Build plugin version 1.7

The Clang Scan-Build plugin release 1.7 adds an option to specify the
directory where the generated reports are expected.

Without this change, builds that have a Scan-Build-Publisher configured
fail with a NPE because the 'reportFolderName' element is missing in
the config.

Reference: https://issues.jenkins-ci.org/browse/JENKINS-18941

Change-Id: Ide3627a58198c81744f9c360f45929a23dce5d23