Updated openstack/openstack Project: openstack/nova-specs  8c717ca94e5420e3304e4473b74363ac371ac591
Return 400 When Bad Status Values are Received

A bug was found where passing an invalid status to the server listing
API would return an empty list for a regular user, but would raise an
exception and return a 500 error for admin users. In the discussions
about that bug it was felt that while fixing the exception was certainly
correct, the list of valid status values is small and well-defined, so
returning an empty list was not the desired behavior; instead, the user
should get a 400 Bad Request instead.

Blueprint response-for-invalid-status

APIImpact

Change-Id: Id5a9b6708b39530ca6f53e9c576305d01f211c78