Updated openstack/openstack Project: openstack-infra/tripleo-ci  fd337aac59e46d8c1c902e302c643897d9307e63
pingtest: don't show error if no images needs to be cleaned-up

When running pingtest, we clean-up old images created by a previous
pingtest but don't fail if no image does exist. Though we see this
message:
openstack image delete: error: too few arguments

It's because no image does exist, so ID is empty, we don't need to show
it in logs, it's confusing and useless. Let's redirect it's stderr to
null.

Change-Id: Icb8e6c0036b46b607e6c6d671e91c3a3c59a3976