Updated openstack/openstack Project: openstack/keystone  24660d7841a059e68c57131a21337cf067d87b94
Verify domain_id when get_domain is being called

When create user using API it is possible to use a domain_id which
does match the created domain_id since mysql per default is not
case sensitive and returns the domain_id to be valid. In e.g.
liberty this breaks cli keystone v2 user list actions when a user
with a DEFaULt domain has been created.

With this change the domain_id is being validated with what provided
with the API call in get_domain.

Change-Id: I028b2add3067e6fb9aa3f33eb8fe10d8ebace006
Closes-Bug: #1594284