Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  6c9d8ccb0cc63cf2ad7f0c4659c731e2fcbb0988
Properties: fix copyartifact permissions with several projects

The copyartifacts property on Jenkins requires that the config XML
provide an string element for each allowed project. On JJB the
documentation says that the projects needed only to be comma-separated,
however JJB is not doing the job separation.

Parsing the projects string, spliting it by commas and then adding it to
a different string element solves this.

Change-Id: I59d8b2c070f28df4cfa9da0ab58f2c57e4243384
Signed-off-by: Lucas Dutra Nunes <ldnunes@ossystems.com.br>