Updated openstack/openstack Project: openstack/python-novaclient  fddca7dda6fff4f36d11bfd831a5603137089599
Raise an exception in v2.client for direct_use

The direct use of novaclient.v2.client.Client was deprecated long time ago
(2.29.0 release of novaclient, which relates to Liberty release of OpenStack).

It is time to move from warning message to an exception. It will allow us to
change novaclient.v2.client.Client interface and keep compatible stuff to
novaclient.client.Client.

Change-Id: I6e6c6eebcf7992d553a2f85f57b8c2ede0cc8311
Related-Bug: #1493576