Updated openstack/openstack Project: openstack/nova  3e48d05d2174824c334ba2f6ed0476c4171277e7
Extract port create from allocate_for_instance

This pushes the port create loop into a helper function, so we are
slowly pulling apart the create and update stages of
allocate_for_instance.

This change also adds unit tests for that function, isolated from the
rest of the allocate_for_instance logic.

Because we now call all port creates, before doing the port updates,
some of the existing unit tests had to be updated.

blueprint prep-for-network-aware-scheduling

Change-Id: Ia12aa98296b96c1948b162efd1b3a6b717dd3ade