Updated openstack/openstack Project: openstack/tempest  0d6c4c2163f2c7657f655b074b8254d30d12c0d0
Add support for specifying a config file to tempest run

This commit adds a new cli option --config-file for specifying where
the tempest config file lives on disk. This adds the final execution
mode to tempest and enables running tempest from anywhere including
outside of a workspace or tempest repo. It does this making a temporary
.testr.conf on disk (which it cleans up) and a .testrepository (which
stays for history).

Change-Id: I74ae79147244d60bd0ddd9cc460900f3ab69f3f0