Updated openstack/openstack Project: openstack/keystonemiddleware  6a5ef484f2c1767fc332461eb13b3c44fa799c59
Refactor audit tests to use create_middleware

The auth_token tests have a good layout for using create_middleware and
create_simple_middleware that makes it easy for tests to create
middleware for testing as required rather than having a self.middleware
object that gets overriden a lot.

Extract this create_middleware into a base class and implement it in
audit so the patterns are similar.

Change-Id: I2f050eef1684c8046f94dc2b88b4c97a56ea9cd8