Updated openstack/openstack Project: openstack/keystonemiddleware  98b639aeb7c2721ec3d4a09e695d7732541b8322
Use oslo_messaging conf fixture

Oslo_messaging mocking in audit tests is a bit odd because middleware
creation must be run multiple times so that the oslo_messaging opts are
registered before the config fixture tries to set it.

If we use the oslo_messaging conf fixture then it handles all the option
registration for us correctly and removes the need for us to manually
patch get_transport.

Change-Id: Iaea2e66e9418a0426d73e1c4554c56be262a7779