Updated openstack/openstack Project: openstack/tripleo-heat-templates  043d71d5ce6a1d7c3fc3a8da78ac4ea5622bb6b9
Add Rabbit/epmd port to firewall

This commit adds the epmd port 4369 to the firewall configuration for
the service rabbit. This is necessary for having HA setups working,
since without this port the rabbitmq cloned resource starts only on one
node and the others are not able to complete the rabbit cluster
creation.

Change-Id: Iae042dd60a578e158b75539dc3998fc40185b343