Updated openstack/openstack Project: openstack/manila  5b11b87dc732056170861f0e64e044cd7f08e172
Fix snapshot manage Tempest test

In the process of managing a snapshot, drivers may want to
rename the snapshot. This could warrant updating the
provider_location as well.

The tempest tests for snapshot manage should not expect that
the provider location will remain unchanged after the driver
has performed the manage operation.

Change-Id: Ib07ba0c55ef4cb3da8f4ba5b0f9c1d4340c3f0e1
Closes-Bug: #1598287