Updated openstack/openstack Project: openstack/nova  60ef8434cb47ed09b0587e37f0a72a6ba3d9a549
Fix NoMoreNetworks functional test traces

A lot of the functional tests pass but spew stacktraces
with NoMoreNetworks on allocate_for_instance.

We should stub out the network API for the tests that create
servers because they don't populate networks in the db and
the networks aren't relevant to what is being tested.

Co-Authored-By: melanie witt <melwittt@gmail.com>

Change-Id: Idb0ae51d1773a6fd14f4abc30c5186a7e3389a73