Updated openstack/openstack Project: openstack/python-heatclient  d7d76485854c10afba08fdf11267e1a211e7298f
base.Resource not define __ne__() built-in function

Class base.Resource defines __eq__() built-in function, but does
not define __ne__() built-in function, so self.assertEqual works
but self.assertNotEqual does not work at all in this test case in
python2. This patch fixes it by defining __eq__() built-in function
of class base.Resource.

Change-Id: I13ed9ddd6ae6ef074232b33695916bc83930dcf0
Closes-Bug: #1586268