Updated openstack/openstack Project: openstack/heat  0d0651a88c1517d833e94ef5bd1d79a609ee6802
Add default release names for template versions

Release version names are easier for template authors to remember than
seemingly arbitrary dates. It should also make it clearer to new Heat
users that heat_template_version is not like a BIND zone file version
(which would be incremented with each version of the template).

Change-Id: Iae9b6a676702b2a115b18fdfb1ab2afbe5faacde
blueprint support-alias-for-heat-template-version-names