Updated openstack/openstack Project: openstack/cinder  b81f4f2f949c60f712f63cb5d5e5572136bbb080
Add logging when filtering returns nothing

When hosts are filtered for a request, and no host passes all the
filters, operators can only see the last filter run; they have no
visibility into which filter removed which host. This attempts to
improve this situation by tracking the hosts remaining after each
filter is run and the number of hosts removed. If no host is left,
that information along with volume ID of the request would be logged.

Since that can be a lot of information, it will call LOG.info with
just the counts for all but the final filter, which will contain
the last deleted host names, LOG.debug will output the full record
of each filter and the hosts remaining after each step.

This will better enable operators to determine why a particular host
was not selected.

Change-Id: I3baf75a4a4430c55b30dde1cae86fa5e195ec6d3