Updated openstack/openstack Project: openstack-dev/devstack  8e64c478ebdfa76c80567a42ca52c9149ee0e52b
Tempest: add a Keystone reseller feature flag

A new tempest test is being added in https://review.openstack.org/#/c/285541/
but it is not supported in the Kilo and Liberty branches. This patch
turns on this feature flag at Devstacks's side.

According to tempest policies, this patch must be merged first so the
test can actually run.

Change-Id: I52458a0b36e1dba233667311b35f6c3931e2e66c
Depends-On: Ie69dae09c2b42e825e9d51abf158fc14788387d1