Updated openstack/openstack Project: openstack/cinder  a1105e3abd3057274fcd248e9bb2311b9dc81d4a
EQLX: Extend of snapshot or clone failure

When creating a volume from a snapshot or cloning an existing
volume the driver first creates a new volume of the original
size, then extends the volume if needed to meet the requested
size. The extend call in these operations were passing the
wrong arguments into the extend_volume call, resulting in an
exception being thrown.

Change-Id: I503543fc45a0607d295a2ebe3602436857fc8ff8
Closes-bug: #1598148