Updated openstack/openstack Project: openstack-infra/project-config  648c1408f2d72976a4229fb3c3dc1a8bd8cebc06
Make ovn-native-l3 check job voting

The ovn-native-l3 check job was created as non-voting,
and is now stable with basically the same (low) failure
rate as the ovn voting job.  This should be made voting
and added to the gate since the default implementation of
ovn uses native l3 while the existing voting job uses the
neutron legacy q-l3 agent.

Change-Id: I3e1978b97711e406e34513d8cb7b644f2a361caa