Updated openstack/openstack Project: openstack/tempest  d2f46265c1f0000497df97bd464c181aa629151f
Update quickstart for revised temepst init setup

This commit updates the quickstart to reflect reality when using
tempest init. The quickstart documented our original ideal expectation
of how things worked, but because of python packaging limitations we
had to adjust it's behavior. However, the quickstart was never updated
to reflect this change.

Change-Id: Icbc0c005a344c25b7ab7d3fd059a9401747eca3f
Closes-Bug: #1598882