Updated openstack/openstack Project: openstack/oslo.db  24b8812a74979dd6313564d902b28eff562415ba
Catch empty value DBDuplicate errors

We can have integrity constraint violations when the
value is an empty string. The previous regex for converting
these into DBDuplicate exceptions was not considering this
case.

This adjusts the regex to a 0 or many instead of a 1 or many
match for the value and adds a case to a test.

Change-Id: Ibf4686f92fcffb8e70ec2735ab878e8849ed37b1
Closes-Bug: #1599086