Updated openstack/openstack Project: openstack/openstackdocstheme  13bd97688aa51d6b3a292f0c97b7f1e3ea7cb120
Allow the bug title to be customisable

Some projects using openstackdocstheme may wish to customise
the default bug title to be better suited to their needs.

This patch implements the ability to do so.

To define it, add the variable to html_context in the Sphinx
conf.py.

For example:
  html_context = {"bug_title": 'Documentation bug', ...}

Change-Id: Id62debd50fe0499ace809dde626cd8a357dbf933