Updated openstack/openstack Project: openstack/oslo.messaging  58ad758dc21e2ebbd259e23a39cd7c9efea5f00f
Remove rabbitmq max_retries

It was never clear was the purpose of this option.

For the RPC API point of view raising MessageDeliveryFailure is kind
of unexpected. For the Notification API, this is weird that when
application put retry=-1, the lib stop to retry only with rabbit.

So this change deprecates this option for removal.

Change-Id: I4ac3a062426418276c96e1745d1c96a8525c36da