Updated openstack/openstack Project: openstack-infra/project-config  49b568b9877d4472d0576cedd5a7bb982bec777c
Bump VM memory for latest stable CoreOS IPA

IPA is currently running an extremely old CoreOS. New CoreOS ramdisks
require more RAM. Given a CoreOS ramdisk is only built and run for IPA
jobs, bump memory for virtual machine from 1024M to 1280M. This has
been tested as sufficient locally.

It was determined that the stable CoreOS image can boot with no less
than 1171M of RAM. The next multiple of 128 is 1280M. This leaves 109M
of free memory. This should be enough for deploying disk images under
100M.

The limit is also bumped for old branches as this fix is critical
enough to be backported, and to make it consistent.

Change-Id: Ia49daedac0a972769698f04b5e358ac17435b24d
Partial-bug: #1581673