Updated openstack/openstack Project: openstack/manila  443b61f668d4afaa956748924f054a6fbaca4a21
HPE3PAR: Fix filestore quota decrement

HPE3PAR driver keeps a filestore fsquota to limit the space
available to all the nested shares. This fsquota should not
be decremented when read-only shares are removed. In this case
3PAR read-only shares are only used when the files exist in a
snapshot or in a read-write share. So the read-only share
itself does not hold any space. If a read-write share is
removed, then the decrement happens.

Change-Id: I1f006ce7e6cb2473f11b1505c1b7d4946aa5e1fc
Closes-Bug: #1599618