Updated openstack/openstack Project: openstack/heat  150a013597da2605be34a3971430283fc25946da
Fail the test run if image download fails

When we run the integration tests, we download the test image that we
need and upload it on the fly. If the download is aborted because of
some network problem, the image is incorrect, but we still try to run
the tests and boot servers with it. We should fail early instead.

Change-Id: I8797440086206749c9098dcd42b84d95426fbf3e
Closes-Bug: #1534026