Updated openstack/openstack Project: openstack/diskimage-builder  be184fa254fa084afb1d9ff3ba95c877f51ae47b
Clear up "already provided" message

I got quite confused what this was trying to tell me at first.  It's
saying that you requested an element but another element already
provides that element, so we don't know which one to choose.

To help clarify the situation to the user, keep track of what is
providing elements so we can describe where the conflict came from.

Change-Id: Ie7471ac900a8cbee5684c928badd1b8ce6d3e3cf