Updated openstack/openstack Project: openstack/horizon  8850cdda2432f2f28331d4ec8393a8cc33ff5ae8
Modify hz-cell to use hz-field

hz-cell doesn't use params, which causes lots of problems when trying to
render changes to data...since it evaluates only once, reading what's assumed
to be on scope.

This patch adds 'table,' 'column,' and 'item' params so they may be better
watched by the directive.

hz-field does use params and thus is useful in many situations outside
of hz-cell.  This patch provides hz-field and lets hz-cell use it so
all output logic is shared.

hz-field also: a) accepts 'values' as a column configuration as we discussed
in the last mid-cycle, so codes, etc. may be supplied without use of a filter;
and b) accepts 'urlFunction' which translates the given item to a url which is
linked on the outputted name (rather than having to write a template for a
common link function).

Follow-on patches will demonstrate use.

Change-Id: I0835a90e61d0e708233da795964595f88616388c
Partially-Implements: blueprint angular-registry